+++ include/jemalloc/internal/jemalloc_internal_includes.h
@@ -51,7 +51,6 @@
 #include "jemalloc/internal/prof_structs.h"
 #include "jemalloc/internal/arena_structs.h"
 #include "jemalloc/internal/tcache_structs.h"
-#include "jemalloc/internal/background_thread_structs.h"
 
 /******************************************************************************/
 /* EXTERNS */
@@ -59,11 +58,11 @@
 
 #include "jemalloc/internal/jemalloc_internal_externs.h"
 #include "jemalloc/internal/arena_externs.h"
-#include "jemalloc/internal/large_externs.h"
+#include "jemalloc/internal/large.h"
 #include "jemalloc/internal/tcache_externs.h"
 #include "jemalloc/internal/malloc_dispatch.h"
 #include "jemalloc/internal/prof_externs.h"
-#include "jemalloc/internal/background_thread_externs.h"
+#include "jemalloc/internal/background_thread.h"
 
 /******************************************************************************/
 /* INLINES */
