+++ include/jemalloc/internal/malloc_dispatch_inlines.h
@@ -9,7 +9,7 @@
 #include "jemalloc/internal/emap.h"
 #include "jemalloc/internal/jemalloc_internal_inlines_b.h"
 #include "jemalloc/internal/jemalloc_internal_types.h"
-#include "jemalloc/internal/large_externs.h"
+#include "jemalloc/internal/large.h"
 #include "jemalloc/internal/malloc_dispatch.h"
 #include "jemalloc/internal/safety_check.h"
 #include "jemalloc/internal/sc.h"
